Add extra `serializePerson` method for new tests
(In `PersonUtil`)

Previous methods only took a `Set` (to be serialized into a JSON array
of strings) for the tags.

This new method also takes a `String` as the raw representation of the
`tagged` value so that we can test serialization of invalid `Person`
serializers where the value of the `tagged` field is not a JSON array.